Added closing </script> tag to index.html
This missing tag prevented the javascript from executing. It worked
in local environment (using live-server), because live-server injects
its own code to the end of HTML body, and that apparently closes the
script tag. When tested using `python3 -m http.server 8000`, javascript
failed to execute, as noticed on the openflowy.com website.

So, we need a test that doesn't use live-server, and instead tries to
provide a behaviour as close as possible to the production environment.