Added code to programmatically create Tasks
This is a large commit, result of over 24 hours of work, and is
difficult to split into smaller, individual commits. Following describes
the various changes in this commit.

In README,
- Describe the major features concisely
- Updated the 'Clone Project' and Publish sections, to account for the
  slightly changed development and deployent steps.
- Added Debug section to show how to debug the code

In main.ts, created the Task class to represent Task objects in memory.
And, created a demo of the app (list of Tasks) that doubles as the list
of features of the app.

In package.json,
- Added runtime dependency on uuid package, needed for generating UUIDv4
- Added dev-dependencies for
  - @types/uuid, TypeScript type-definitions for the UUID package
  - Webpack, Webpack CLI, webpack-dev-server, and ts-loader for aiding
    in bundling and development
  - comment-json package to parse tsconfig and use its values in Webpack
    configuration
  - Removed live-server package, as we now use Webpack for that purpose

The Webpack related changes are based on Webpack's TypeScript guide at
https://webpack.js.org/guides/typescript/

Also in package.json,
- Updated the various entries under the "scripts" attribute to replace
  live-server with equivalent Webpack commands
- Fixed the commands in "scripts" to use && operator as much as possible
- Added plenty of comments as named attributes
- Renamed property "publish-build-output" to "publish-web-output"

Updated index.html to use bundle.js produced by Webpack.

Updated parameter values in tsconfig.json to enable some critical
checks, and to make it emit development friendly bundle output, with
comments and source-maps.

In .gitignore, ignore the contents of directory created by IDE VSCode.